requirements.txt
aiohttp
asgiref
attrs
backoff
certifi
cffi
channels
charset-normalizer
colorama
cryptography
django==4.2
django-admin-soft-dashboard
django-allauth
django-axes
django-cors-headers
django-filter
django-formtools
django-ipware
django-otp
django-phonenumber-field
django-two-factor-auth
djangorestframework
mysqlclient
fonttools
frozenlist
gunicorn
idna
incremental
joblib
kiwisolver
matplotlib
multidict
numpy
opencage
packaging
pandas
phonenumbers
pillow
plotly
pyasn1
pyasn1_modules
pycparser
PyJWT
python-dateutil
python-dotenv
python3-openid
pytz
qrcode
requests
requests-oauthlib
service-identity
six
social-auth-app-django
social-auth-core
sqlparse
tqdm
Twisted
txaio
types-python-dateutil
typing_extensions
tzdata
Unidecode
urllib3
whitenoise
yarl
zope.interface
